fix: bg color items search-bar
