--- ui/components/popover-header/popover.ts
@@ -1,1 +1,0 @@
-export * from './popover-header.component'
